+++ src/OpenLoco/Vehicles/VehicleManager.cpp
@@ -4,10 +4,16 @@
 #include "../Entities/EntityManager.h"
 #include "../GameCommands/GameCommands.h"
 #include "../GameState.h"
+#include "../Graphics/ImageIds.h"
+#include "../Input.h"
 #include "../Interop/Interop.hpp"
 #include "../MessageManager.h"
+#include "../Objects/CargoObject.h"
+#include "../StationManager.h"
 #include "../Ui/WindowManager.h"
+#include "Orders.h"
 #include "Vehicle.h"
+#include <sstream>
 
 using namespace OpenLoco::Interop;
 
@@ -226,6 +232,10 @@ namespace OpenLoco::Vehicles::RoutingManager
 
 namespace OpenLoco::Vehicles::OrderManager
 {
+    // TODO: Make this a fixed vector of size 63 (kMaxNumOrderPerVehicle) no need for it to be dynamic
+    std::vector<UnkF2494A> _displayFrames;
+    const std::vector<UnkF2494A>& displayFrames() { return _displayFrames; }
+
     static auto& orders() { return getGameState().orders; }
     static auto& numOrders() { return getGameState().numOrders; }
 
@@ -254,5 +264,164 @@ namespace OpenLoco::Vehicles::OrderManager
         std::rotate(&orders()[offset], &orders()[offset + size], &orders()[numOrders()]);
 
         numOrders() -= head->sizeOfOrderTable;
+    }
+
+    // 0x00470B76
+    std::pair<Map::Pos3, std::string> generateOrderUiStringAndLoc(uint32_t orderOffset, uint8_t orderNum)
+    {
+        std::stringstream ss;
+        ss << ControlCodes::inlineSpriteStr;
+
+        auto imageId = Gfx::recolour(ImageIds::getNumberCircle(orderNum), Colour::white);
+        ss.write(reinterpret_cast<const char*>(&imageId), 4);
+        OrderRingView orderRing(orderOffset);
+        auto order = orderRing.begin();
+        Map::Pos3 pos{};
+
+        switch (order->getType())
+        {
+            case Vehicles::OrderType::StopAt:
+            {
+                auto* stopAt = order->as<Vehicles::OrderStopAt>();
+                // 0x00470B7D
+                auto* station = StationManager::get(stopAt->getStation());
+                pos = Map::Pos3{ station->x, station->y, station->z } + Map::Pos3{ 0, 0, 30 };
+                ss << ControlCodes::Colour::white;
+                for (auto nextOrder = order + 1; nextOrder != order; ++nextOrder)
+                {
+                    if (!nextOrder->hasFlag(Vehicles::OrderFlags::HasCargo))
+                    {
+                        break;
+                    }
+                    auto* unload = nextOrder->as<Vehicles::OrderUnloadAll>();
+                    auto* waitFor = nextOrder->as<Vehicles::OrderWaitFor>();
+                    const CargoObject* cargoObj = nullptr;
+                    if (unload != nullptr)
+                    {
+                        ss << " - ";
+                        cargoObj = ObjectManager::get<CargoObject>(unload->getCargo());
+                    }
+                    else if (waitFor != nullptr)
+                    {
+                        ss << " + ";
+                        cargoObj = ObjectManager::get<CargoObject>(waitFor->getCargo());
+                    }
+                    else
+                    {
+                        // Not possible
+                        break;
+                    }
+                    ss << ControlCodes::inlineSpriteStr;
+                    ss.write(reinterpret_cast<const char*>(&cargoObj->unitInlineSprite), 4);
+                }
+                break;
+            }
+            case Vehicles::OrderType::RouteThrough:
+            {
+
+                // 0x00470C25
+                auto* station = StationManager::get(order->as<Vehicles::OrderRouteThrough>()->getStation());
+                pos = Map::Pos3{ station->x, station->y, station->z } + Map::Pos3{ 0, 0, 30 };
+                break;
+            }
+            case Vehicles::OrderType::RouteWaypoint:
+                // 0x00470C6F
+                pos = order->as<Vehicles::OrderRouteWaypoint>()->getWaypoint() + Map::Pos3{ 16, 16, 8 };
+                break;
+            case Vehicles::OrderType::End:
+            case Vehicles::OrderType::UnloadAll:
+            case Vehicles::OrderType::WaitFor:
+                return { {}, {} };
+        }
+        ss << std::endl;
+        return { pos, ss.str() };
+    }
+
+    // 0x00470824
+    void sub_470824(Vehicles::VehicleHead* head)
+    {
+        Input::setMapSelectionFlags(Input::MapSelectionFlags::unk_04);
+        Gfx::invalidateScreen();
+        _displayFrames.clear();
+        auto orders = Vehicles::OrderRingView(head->orderTableOffset);
+        uint8_t i = 0;
+        for (auto& order : orders)
+        {
+            if (!order.hasFlag(Vehicles::OrderFlags::HasNumber))
+            {
+                continue;
+            }
+            UnkF2494A newFrame;
+            newFrame.orderOffset = order.getOffset();
+
+            auto lineNumber = 0;
+            auto* stationOrder = order.as<Vehicles::OrderStation>();
+            auto* waypointOrder = order.as<Vehicles::OrderRouteWaypoint>();
+            if (stationOrder != nullptr)
+            {
+                lineNumber++; // station labels start on line 0 so add 1 for the number
+
+                const auto stationId = stationOrder->getStation();
+                for (auto innerOrder = orders.begin(); innerOrder->getOffset() != order.getOffset(); ++innerOrder)
+                {
+                    auto* innerStationOrder = innerOrder->as<Vehicles::OrderStation>();
+                    if (innerStationOrder == nullptr)
+                    {
+                        continue;
+                    }
+                    if (stationId == innerStationOrder->getStation())
+                    {
+                        lineNumber++;
+                    }
+                }
+            }
+            else if (waypointOrder != nullptr)
+            {
+                const auto rawOrder = order.getRaw();
+                for (auto innerOrder = orders.begin(); innerOrder->getOffset() != order.getOffset(); ++innerOrder)
+                {
+                    if (rawOrder == innerOrder->getRaw())
+                    {
+                        lineNumber++;
+                    }
+                }
+            }
+
+            // For some reason save only a byte when this could in theory be larger
+            newFrame.lineNumber = static_cast<uint8_t>(lineNumber);
+            _displayFrames.push_back(newFrame);
+            i++;
+        }
+
+        i = 1;
+        for (auto& unk : _displayFrames)
+        {
+            auto order = Vehicles::OrderRingView(unk.orderOffset, 0).begin();
+            if (!order->hasFlag(Vehicles::OrderFlags::HasNumber))
+            {
+                continue;
+            }
+
+            auto [loc, str] = generateOrderUiStringAndLoc(order->getOffset(), i);
+            const auto pos = Map::gameToScreen(loc, Ui::WindowManager::getCurrentRotation());
+            auto stringWidth = Gfx::getStringWidth(str.c_str());
+            for (auto zoom = 0; zoom < 4; ++zoom)
+            {
+                // The first line of the label will always be at the centre
+                // of the station/waypoint. This works out where the subsequent
+                // lines of the label will end up.
+                auto width = (stringWidth + 3) << zoom;
+                auto numberHeight = (unk.lineNumber * 10 /* lineHeight TODO make same as Windows::Vehicle.cpp lineHeight */) << zoom;
+                auto firstLineHeight = 11 << zoom;
+                auto midX = width / 2;
+                auto midFirstLineY = firstLineHeight / 2;
+
+                unk.frame.left[zoom] = (pos.x - midX) >> zoom;
+                unk.frame.right[zoom] = (pos.x + midX) >> zoom;
+                unk.frame.top[zoom] = (pos.y - midFirstLineY + numberHeight) >> zoom;
+                unk.frame.bottom[zoom] = (pos.y + midFirstLineY + numberHeight) >> zoom;
+            }
+            i++;
+        }
     }
 }
