+++ src/OpenLoco/Windows/Vehicle.cpp
@@ -33,6 +33,7 @@
 #include "../Ui/ScrollView.h"
 #include "../Ui/WindowManager.h"
 #include "../Vehicles/Orders.h"
+#include "../Vehicles/VehicleManager.h"
 #include "../ViewportManager.h"
 #include "../Widget.h"
 #include <map>
@@ -2359,187 +2360,10 @@ namespace OpenLoco::Ui::Windows::Vehicle
     namespace Route
     {
         static loco_global<uint8_t, 0x00113646A> _113646A;
-#pragma pack(push, 1)
-        struct UnkF2494A
-        {
-            uint32_t orderOffset; // 0x0
-            LabelFrame frame;     // 0x4
-            uint8_t lineNumber;   // 0x24
-            uint8_t pad_25;
-        };
-        static_assert(sizeof(UnkF2494A) == 0x26);
-#pragma pack(pop)
-
-        static loco_global<UnkF2494A[63], 0x00F2494A> _F2494A;
 
         static Vehicles::OrderRingView getOrderTable(const Vehicles::VehicleHead* const head)
         {
             return Vehicles::OrderRingView(head->orderTableOffset);
-        }
-
-        // 0x00470B76
-        static std::pair<Map::Pos3, std::string> generateOrderUiStringAndLoc(const Vehicles::OrderRingView::Iterator& order, uint8_t orderNum)
-        {
-            std::stringstream ss;
-            ss << ControlCodes::inlineSpriteStr;
-
-            auto imageId = Gfx::recolour(ImageIds::getNumberCircle(orderNum), Colour::white);
-            ss.write(reinterpret_cast<const char*>(&imageId), 4);
-
-            Map::Pos3 pos{};
-
-            switch (order->getType())
-            {
-                case Vehicles::OrderType::StopAt:
-                {
-                    auto* stopAt = order->as<Vehicles::OrderStopAt>();
-                    // 0x00470B7D
-                    auto* station = StationManager::get(stopAt->getStation());
-                    pos = Map::Pos3{ station->x, station->y, station->z } + Map::Pos3{ 0, 0, 30 };
-                    ss << ControlCodes::Colour::white;
-                    for (auto nextOrder = order + 1; nextOrder != order; ++nextOrder)
-                    {
-                        if (!nextOrder->hasFlag(Vehicles::OrderFlags::HasCargo))
-                        {
-                            break;
-                        }
-                        auto* unload = nextOrder->as<Vehicles::OrderUnloadAll>();
-                        auto* waitFor = nextOrder->as<Vehicles::OrderWaitFor>();
-                        const CargoObject* cargoObj = nullptr;
-                        if (unload != nullptr)
-                        {
-                            ss << " - ";
-                            cargoObj = ObjectManager::get<CargoObject>(unload->getCargo());
-                        }
-                        else if (waitFor != nullptr)
-                        {
-                            ss << " + ";
-                            cargoObj = ObjectManager::get<CargoObject>(waitFor->getCargo());
-                        }
-                        else
-                        {
-                            // Not possible
-                            break;
-                        }
-                        ss << ControlCodes::inlineSpriteStr;
-                        ss.write(reinterpret_cast<const char*>(&cargoObj->unitInlineSprite), 4);
-                    }
-                    break;
-                }
-                case Vehicles::OrderType::RouteThrough:
-                {
-
-                    // 0x00470C25
-                    auto* station = StationManager::get(order->as<Vehicles::OrderRouteThrough>()->getStation());
-                    pos = Map::Pos3{ station->x, station->y, station->z } + Map::Pos3{ 0, 0, 30 };
-                    break;
-                }
-                case Vehicles::OrderType::RouteWaypoint:
-                    // 0x00470C6F
-                    pos = order->as<Vehicles::OrderRouteWaypoint>()->getWaypoint() + Map::Pos3{ 16, 16, 8 };
-                    break;
-                case Vehicles::OrderType::End:
-                case Vehicles::OrderType::UnloadAll:
-                case Vehicles::OrderType::WaitFor:
-                    return { {}, {} };
-            }
-            ss << std::endl;
-            return { pos, ss.str() };
-        }
-
-        // 0x00470824
-        static void sub_470824(Vehicles::VehicleHead* head)
-        {
-            Input::setMapSelectionFlags(Input::MapSelectionFlags::unk_04);
-            Gfx::invalidateScreen();
-            for (auto& unk : _F2494A)
-            {
-                unk.orderOffset = static_cast<uint32_t>(-1);
-            }
-            auto orders = getOrderTable(head);
-            uint8_t i = 0;
-            for (auto& order : orders)
-            {
-                if (!order.hasFlag(Vehicles::OrderFlags::HasNumber))
-                {
-                    continue;
-                }
-
-                _F2494A[i].orderOffset = order.getOffset();
-
-                auto lineNumber = 0;
-                auto* stationOrder = order.as<Vehicles::OrderStation>();
-                auto* waypointOrder = order.as<Vehicles::OrderRouteWaypoint>();
-                if (stationOrder != nullptr)
-                {
-                    lineNumber++; // station labels start on line 0 so add 1 for the number
-
-                    const auto stationId = stationOrder->getStation();
-                    for (auto innerOrder = orders.begin(); innerOrder->getOffset() != order.getOffset(); ++innerOrder)
-                    {
-                        auto* innerStationOrder = innerOrder->as<Vehicles::OrderStation>();
-                        if (innerStationOrder == nullptr)
-                        {
-                            continue;
-                        }
-                        if (stationId == innerStationOrder->getStation())
-                        {
-                            lineNumber++;
-                        }
-                    }
-                }
-                else if (waypointOrder != nullptr)
-                {
-                    const auto rawOrder = order.getRaw();
-                    for (auto innerOrder = orders.begin(); innerOrder->getOffset() != order.getOffset(); ++innerOrder)
-                    {
-                        if (rawOrder == innerOrder->getRaw())
-                        {
-                            lineNumber++;
-                        }
-                    }
-                }
-
-                // For some reason save only a byte when this could in theory be larger
-                _F2494A[i].lineNumber = static_cast<uint8_t>(lineNumber);
-                i++;
-            }
-
-            i = 1;
-            for (auto& unk : _F2494A)
-            {
-                if (unk.orderOffset == static_cast<uint32_t>(-1))
-                {
-                    continue;
-                }
-
-                auto order = Vehicles::OrderRingView(unk.orderOffset, 0).begin();
-                if (!order->hasFlag(Vehicles::OrderFlags::HasNumber))
-                {
-                    continue;
-                }
-
-                auto [loc, str] = generateOrderUiStringAndLoc(order, i);
-                const auto pos = gameToScreen(loc, WindowManager::getCurrentRotation());
-                auto stringWidth = Gfx::getStringWidth(str.c_str());
-                for (auto zoom = 0; zoom < 4; ++zoom)
-                {
-                    // The first line of the label will always be at the centre
-                    // of the station/waypoint. This works out where the subsequent
-                    // lines of the label will end up.
-                    auto width = (stringWidth + 3) << zoom;
-                    auto numberHeight = (unk.lineNumber * lineHeight) << zoom;
-                    auto firstLineHeight = 11 << zoom;
-                    auto midX = width / 2;
-                    auto midFirstLineY = firstLineHeight / 2;
-
-                    unk.frame.left[zoom] = (pos.x - midX) >> zoom;
-                    unk.frame.right[zoom] = (pos.x + midX) >> zoom;
-                    unk.frame.top[zoom] = (pos.y - midFirstLineY + numberHeight) >> zoom;
-                    unk.frame.bottom[zoom] = (pos.y + midFirstLineY + numberHeight) >> zoom;
-                }
-                i++;
-            }
         }
 
         // 0x004B509B
@@ -2555,7 +2379,7 @@ namespace OpenLoco::Ui::Windows::Vehicle
         {
             GameCommands::setErrorTitle(StringIds::empty);
             GameCommands::do_36(head->id, orderOffset - head->orderTableOffset);
-            sub_470824(head);
+            Vehicles::OrderManager::sub_470824(head);
         }
 
         // 0x004B4F6D
@@ -2592,7 +2416,7 @@ namespace OpenLoco::Ui::Windows::Vehicle
         {
             GameCommands::setErrorTitle(StringIds::empty);
             auto result = GameCommands::do_75(head->id, orderOffset - head->orderTableOffset);
-            sub_470824(head); // Note order changed check if this matters.
+            Vehicles::OrderManager::sub_470824(head); // Note order changed check if this matters.
             return result != GameCommands::FAILURE;
         }
 
@@ -2601,7 +2425,7 @@ namespace OpenLoco::Ui::Windows::Vehicle
         {
             GameCommands::setErrorTitle(StringIds::empty);
             auto result = GameCommands::do_76(head->id, orderOffset - head->orderTableOffset);
-            sub_470824(head); // Note order changed check if this matters.
+            Vehicles::OrderManager::sub_470824(head); // Note order changed check if this matters.
             return result != GameCommands::FAILURE;
         }
 
@@ -2799,7 +2623,7 @@ namespace OpenLoco::Ui::Windows::Vehicle
             GameCommands::setErrorTitle(StringIds::orders_cant_insert);
             auto previousSize = head->sizeOfOrderTable;
             GameCommands::do_35(head->id, order.getRaw(), chosenOffset);
-            sub_470824(head);
+            Vehicles::OrderManager::sub_470824(head);
             if (head->sizeOfOrderTable == previousSize)
             {
                 return;
@@ -2862,7 +2686,7 @@ namespace OpenLoco::Ui::Windows::Vehicle
                 if (Input::toolSet(&self, widx::tool, CursorId::crosshair))
                 {
                     self.invalidate();
-                    sub_470824(head);
+                    Vehicles::OrderManager::sub_470824(head);
                 }
             }
         }
@@ -3145,7 +2969,7 @@ namespace OpenLoco::Ui::Windows::Vehicle
             {
                 return;
             }
-            sub_470824(head);
+            Vehicles::OrderManager::sub_470824(head);
         }
 
         // 0x004B468C
